refactor(routes): rename routes/devices to routes/devices-sessions
