+++ lib/routes/index.js
@@ -34,7 +34,7 @@ module.exports = function (
     push
   )
   const devicesImpl = require('../devices')(log, db, push)
-  const devices = require('./devices')(log, db, config, customs, push, devicesImpl)
+  const devicesSessions = require('./devices-sessions')(log, db, config, customs, push, devicesImpl)
   const password = require('./password')(
     log,
     db,
@@ -62,7 +62,7 @@ module.exports = function (
 
   const v1Routes = [].concat(
     account,
-    devices,
+    devicesSessions,
     password,
     session,
     signinCodes,
